2020-09-05 백엔드 변경 - 사용자 확장 ID 버그 수정

+++ MAMBrowser/DTO/DTO_RESULT_PAGE_LIST.cs
@@ -10,6 +10,6 @@ namespace MAMBrowser.DTO
         public IList<T> Data { get; set; }
         public int RowPerPage { get; set; }     //페이지당 행 수
         public int SelectPage { get; set; }     //현재 페이지
-        public int TotalRowCount { get; set; } = 0;    //전체 행 수
+        public long TotalRowCount { get; set; } = 0;    //전체 행 수
     }
 }
